nexum/backend/pyproject.toml
[project]
name = "nexum"
version = "4.1.0"
description = "Calibrated Acoustic Intelligence Engine with Multimodal Social Knowledge Graph"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "yt-dlp>=2024.8.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "torch>=2.3.0",
    "transformers>=4.44.0",
    "sentence-transformers>=3.0.0",
    "open-clip-torch>=2.26.0",
    "faster-whisper>=1.0.0",
    "easyocr>=1.7.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.4.0",
    "opencv-python-headless>=4.10.0",
    "numpy>=1.26.0",
    "scipy>=1.14.0",
    "scikit-learn>=1.5.0",
    "qdrant-client>=1.11.0",
    "minio>=7.2.0",
    "pyyaml>=6.0.0",
    "python-multipart>=0.0.9",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "structlog>=24.4.0",
    "prometheus-client>=0.20.0",
    "tenacity>=8.5.0",
    "scenedetect[opencv]>=0.6.4",
    # ── Social Knowledge Graph (v3) ──
    "neo4j>=5.20.0",
    "spacy>=3.7.0",
    "langdetect>=1.0.9",
    "websockets>=12.0",
    # ── Acoustic Intelligence (v3.1) ──
    "librosa>=0.10.0",
    "panns-inference>=0.1.1",
    "soundfile>=0.12.0",
    # ── Calibrated Intelligence (v4) ──
    "joblib>=1.4.0",          # calibration model persistence
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"
